% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_functions.R
\name{makepredf}
\alias{makepredf}
\title{Make dataframe for visibility model prediction.}
\usage{
makepredf(covdf, depths)
}
\arguments{
\item{covdf}{A single-row data frame of explanatory variables for \code{model}.}

\item{depths}{The depths to appear in the \code{depth} column of the output
dataframes.}
}
\value{
If the input dataframe has a column named \code{pos}, a single
dataframe is output, with \code{depths} in its \code{depth} column, and the
other columns being identical copies of teh single-row input dataframe
\code{covdf}.

If the input dataframe does not a column named \code{pos}, then a list is
output, containing three dataframes like that for the case above, each with
a column \code{pos} that contains either "Centre", "Near" or "Far". These
dataframes are named $centre, $near and $far. This is so that we can predict
for each of these three levels of \code{pos} and then marginalise the
predictions over \code{pos}.
}
\description{
Make dataframe with \code{depth} column containing \code{depths}, from
a single-row dataframe (\code{covdf}) that does not contain a \code{depth}
column.
}
\examples{
data("eg_viscov")
data("vismodel")
p = plotdetfn(vismodel,eg_viscov)

}
