% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_functions.R
\name{plotavail}
\alias{plotavail}
\title{Plots an availability function.}
\usage{
plotavail(
  depths,
  tagmodel = NULL,
  covdf = NULL,
  what = "CDF",
  doplot = TRUE,
  ...
)
}
\arguments{
\item{depths}{Depths at which to evaluate the availability function.}

\item{tagmodel}{A gam object defining p(depth<=2). If missing then interpolated
values of the CDF of Tielman et al. (2013) are returned.}

\item{covdf}{A single-row data frame of explanatory variables for \code{tagmodel}.
This is ignored if no \code{tagmodel} is passed.}

\item{what}{If "CDF" then the cumulative distribution function is plotted, else
probability density function is plotted.}

\item{doplot}{If TRUE, plotting is done, else only values are calculated (in
which case you need to use an assignment (see examples) to get hold of the
values), else you get nothing from the function.}

\item{...}{Arguments to \code{plot}, other than \code{type}, \code{xlab},
\code{ylab}, which are hardwired.}
}
\value{
Invisibly returns a data frame with columns \code{$depth}, and the
cdf or pdf (called \code{$F0), \code{$F), \code{$f0), or \code{$f),
depending what you asked for}
}
\description{
Plots an availability function.
}
\examples{
data("eg_tagcov")
data("tagmodel")
depths = seq(0,3,length=21)
# To plot and print the cdf (F0) of Tielman et al. (2013):
par(mfrow=c(2,2))
F0 = plotavail(depths,tagmodel=NULL,covdf=NULL,what="CDF",type="l")
F0
# To plot and print the scaled cdf (F) for availability:
F = plotavail(depths,tagmodel=tagmodel,covdf=eg_tagcov[1,],what="CDF",type="l")
F
# To plot and print the pdf (f0) of Tielman et al. (2013):
f0 = plotavail(depths,tagmodel=NULL,covdf=NULL,what="PDF",type="l")
f0
# To plot and print the scaled pdf (f) for availability:
f = plotavail(depths,tagmodel=tagmodel,covdf=eg_tagcov[1,],what="PDF",type="l")
f
}
