% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_functions.R
\name{plotdetfn}
\alias{plotdetfn}
\title{Plots a detection function.}
\usage{
plotdetfn(model, covdf, ndepths = 50, dmax = 3, addCI = TRUE, ...)
}
\arguments{
\item{model}{A gam object defining p(see|depth).}

\item{covdf}{A single-row data frame of explanatory variables for \code{model}.}

\item{ndepths}{The number of depths at which to calculate p(see|depth). Depths
are evenly spaced from 0 to \code{dmax}}

\item{dmax}{The maximum depth to consider}

\item{addCI}{If TRUE 95 percent CI lines are added, UNLESS \code{pseex} returns
NAs for $lcl and $ucl (which it currently does when marginalising over variable pos).}

\item{...}{Arguments to \code{plot}, other than \code{ylim}, \code{xlab},
\code{ylab}, which are hardwired.}
}
\value{
Invisibly returns a list with \code{depth}, \code{p}, \code{lcl}, \code{ucl}.
}
\description{
Plots a detection function.
}
\examples{
data("eg_viscov")
data("vismodel")
plotdetfn(vismodel,eg_viscov)

}
