% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_functions.R
\name{predict2m}
\alias{predict2m}
\title{Predict probability of being in (0m,2m).}
\usage{
predict2m(newdata)
}
\arguments{
\item{newdata}{A data frame containing explanatory variables
\code{diy} (day in year, numeric), and \code{periodfac} (period factor: "0",
"1","2",or "3").}
}
\value{
A vector of probabilities of an animal being in 0m to 2m.
}
\description{
Predicts mean probability of an animal being in 0m to 2m (i.e., at mean of
the individual random effect), using the gam model \code{tagmodel}.
}
\examples{
newdata = data.frame(diy=c(0,100,200,300,365),periodfac=c("0","1","2","3","0"))
predict2m(newdata)

}
