% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_functions.R
\name{psee}
\alias{psee}
\title{Predict detection probabilities.}
\usage{
psee(viscov, tagcov, vismodel, tagmodel)
}
\arguments{
\item{viscov}{A data frame of explanatory variables for \code{vismodel}, each
row giving the covariates associated with one detection.}

\item{tagcov}{A data frame of explanatory variables for \code{tagmodel}, each
row giving the covariates associated with one detection (the same detection as
for the corresponding row of \code{viscov}.}

\item{vismodel}{A gam object defining p(see|depth).}

\item{tagmodel}{A gam object defining p(depth<2).}
}
\value{
A vector of marginal probabilities of detecting animals with visibility
covariates as specified by the rows of \code{viscov}.
}
\description{
Predicts detection probability when a dataframe with one or more rows
containing a sets of visibility covariates is input. The function just
loops through each row calling \code{psee1} with each row.
}
\examples{
data("eg_tagcov")
data("eg_viscov")
data("eg_viscov_withpos")
data("vismodel")
data("tagmodel")
eg_viscov
p = psee(eg_viscov,eg_tagcov, vismodel,tagmodel) # marginalising over pos
eg_viscov_withpos
p_withpos = psee(eg_viscov_withpos,eg_tagcov,vismodel,tagmodel) # marginalising over pos
p;p_withpos

}
