% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_functions.R
\name{psee1}
\alias{psee1}
\title{Predict a single detection probability.}
\usage{
psee1(viscov, tagcov, vismodel, tagmodel)
}
\arguments{
\item{viscov}{A single-row data frame of explanatory variables for \code{vismodel}.}

\item{tagcov}{A single-row data frame of explanatory variables for \code{tagmodel}.}

\item{vismodel}{A gam object defining p(see|depth).}

\item{tagmodel}{A gam object defining p(depth<2).}
}
\value{
The marginal probability of detecting an animal.
}
\description{
Predicts detection probability when a dataframe with one row containing a
single set of visibility covariates is input.
}
\examples{
data("eg_tagcov")
data("eg_viscov")
data("eg_viscov_withpos")
data("vismodel")
data("tagmodel")
eg_viscov
p = psee1(eg_viscov[1,],eg_tagcov[1,], vismodel,tagmodel) # marginalising over pos
eg_viscov_withpos
p_withpos = psee1(eg_viscov_withpos[1,],eg_tagcov[1,],vismodel,tagmodel) # marginalising over pos
p;p_withpos

}
