% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_functions.R
\name{pseex}
\alias{pseex}
\title{P(see|depth) at a range of depths.}
\usage{
pseex(model, covdf, ndepths = 50, dmax = 3, addCI = TRUE)
}
\arguments{
\item{model}{A gam object defining p(see|depth).}

\item{covdf}{A single-row data frame of explanatory variables for \code{model}.}

\item{ndepths}{The number of depths at which to calculate p(see|depth). Depths
are evenly spaced from 0 to \code{dmax}}

\item{dmax}{The maximum depth to consider}

\item{addCI}{If TRUE, lower and upper 95 percent confidence intervals are calculated.}
}
\value{
Returns a list containing the estimated detection probability at each
depth (\code{p}), the lower and upper 95\% CI values at these depths (\code{lcl}
and \code{ucl}), and the depths at which the probabilities were calculated
(\code{depth).
}
\description{
Calculates p(see|depth) at a range of depths.
}
\examples{
data("eg_viscov")
data("vismodel")
p = plotdetfn(vismodel,eg_viscov)

}
